requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
PyJWT==2.8.0
requests==2.31.0
gunicorn==21.2.0
eventlet==0.33.3
python-dotenv==1.0.0
Werkzeug==2.3.7
assemblyai>=0.30.0
flask-socketio==5.3.6
websocket-client==1.6.4
python-magic==0.4.27
